# docs/openapi/marketplace.yaml
openapi: 3.1.0
info:
  title: BMPL Marketplace API
  version: "2.0.0"
  description: >
    Belize Marketplace & Logistics — Phase 2 marketplace endpoints (categories,
    vendors, products, images, variants, inventory, public catalog/search).
    All paths are under the `/api` prefix. Money is integer minor units (BZD cents).
servers:
  - url: https://bmplapi-production.up.railway.app/api
    description: Production
  - url: http://localhost:4000/api
    description: Local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: Validation failed }
              errors:
                type: array
                items:
                  type: object
                  properties: { path: { type: string }, message: { type: string } }
...
    Forbidden: { description: Missing role/permission }
    NotFound: { description: Not found or not owned }
    Conflict: { description: State/uniqueness conflict }
...
paths:
...
  /admin/categories:
    get: { tags: [Admin], summary: List categories, security: [{ cookieAuth: [] }], responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Forbidden" } }, description: "Requires categories.manage" }
    post: { tags: [Admin], summary: Create category, security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK }, "409": { $ref: "#/components/responses/Conflict" } }, description: "Requires categories.manage" }
  /admin/categories/{id}:
    patch: { tags: [Admin], summary: Update category, parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: OK }, "400": { $ref: "#/components/responses/ValidationError" } }, description: "Requires categories.manage" }
    delete: { tags: [Admin], summary: Delete category, parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: OK }, "409": { $ref: "#/components/responses/Conflict" } }, description: "Requires categories.manage" }
...
  /driver/jobs/{id}:
    get: { tags: [Dispatch], summary: "Own job detail (404 for another driver's job)", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/NotFound" } } }
...
  /jobs:
    get: { tags: [Jobs], summary: "Public job search", parameters: [ { name: q, in: query, schema: { type: string } }, { name: category, in: query, schema: { type: string } }, { name: district, in: query, schema: { type: string } }, { name: employmentType, in: query, schema: { type: string } }, { name: workArrangement, in: query, schema: { type: string } }, { name: remote, in: query, schema: { type: boolean } }, { name: salaryMin, in: query, schema: { type: number } }, { name: sort, in: query, schema: { type: string, enum: [relevance, newest, deadline, salary_asc, salary_desc] } }, { name: page, in: query, schema: { type: integer } } ], responses: { "200": { description: "{ total, page, pageSize, items }" } }, description: "PUBLISHED jobs of APPROVED employers only; excludes expired/closed/draft" }
...
  /jobs/{slug}:
    get: { tags: [Jobs], summary: "Public job detail (+ company + related)", parameters: [ { name: slug, in: path, required: true, schema: { type: string } } ], responses: { "200": { description: OK }, "404": { $ref: "#/components/responses/NotFound" } } }
...
  /employer/profile:
    get: { tags: [Jobs], summary: "Own company profile (APPROVED employer)", security: [{ cookieAuth: [] }], responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Forbidden" } } }
    put: { tags: [Jobs], summary: "Upsert company profile", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "200": { description: OK } } }
  /employer/jobs:
    get: { tags: [Jobs], summary: "Own job listings", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } } }
    post: { tags: [Jobs], summary: "Create a draft job", security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK } }, description: "Requires APPROVED EMPLOYER role" }
  /employer/jobs/{id}/submit:
    post: { tags: [Jobs], summary: "Submit a job for moderation", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: SUBMITTED } }, description: "Employer cannot self-publish — admin approves" }
  /employer/applications/{id}/status:
    post: { tags: [Jobs], summary: "Advance an applicant (validated transition)", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK }, "400": { description: "Illegal transition" }, "404": { description: "Not your applicant" } }, description: "body { status, note? }; own jobs only; writes event + audit + notification + system message" }
  /employer/applications/{id}/resume-url:
    get: { tags: [Jobs], summary: "Signed résumé URL (own applicant only)", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: "{ url }" }, "404": { $ref: "#/components/responses/NotFound" } } }
  /employer/applications/{id}/interviews:
    post: { tags: [Jobs], summary: "Schedule an interview", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK } }, description: "body { scheduledAt, mode, location?, notes? }; no external calendar/video" }
  /admin/jobs:
    get: { tags: [Jobs], summary: "Job moderation queue", parameters: [ { name: status, in: query, schema: { type: string } }, { name: reported, in: query, schema: { type: boolean } } ], security: [{ cookieAuth: [] }], responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Forbidden" } }, description: "Requires jobs.read" }
  /admin/jobs/{id}/moderate:
    post: { tags: [Jobs], summary: "Moderate a job", parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], security: [{ cookieAuth: [], csrfToken: [] }], responses: { "201": { description: OK } }, description: "Requires jobs.moderate; body { action (APPROVE|REJECT|REQUEST_INFO|UNPUBLISH|SUSPEND|ARCHIVE), reason? }" }
  /admin/jobs/reports:
    get: { tags: [Jobs], summary: "Job report queue", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } }, description: "Requires jobs.read" }
  /admin/jobs/analytics:
    get: { tags: [Jobs], summary: "Belize Connect admin analytics", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } }, description: "Requires jobs.read; read-only aggregation" }
  /admin/jobs/employers:
    get: { tags: [Jobs], summary: "Employer companies (suspend/restore via /:id/suspend|restore)", security: [{ cookieAuth: [] }], responses: { "200": { description: OK } }, description: "Requires employers.read; approval itself is via the role-application queue" }